Extract ENR's `multiaddrs` field codec to separate module
